LLM-DDS_base/image2text_conversion-main/requirements_learn.txt
absl-py==2.1.0
accelerate==0.33.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
apyori==1.1.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
boto3==1.35.70
botocore==1.35.70
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==2.0.0
chatoperastore==1.2.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
community==1.0.0b1
contourpy==1.2.1
cycler==0.12.1
d2l==0.17.0
datasets==2.16.1
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.30.1
dill==0.3.7
distro==1.9.0
dnspython==2.6.1
einops==0.8.0
email_validator==2.2.0
emoji==2.14.0
et_xmlfile==2.0.0
evaluate==0.4.2
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastjsonschema==2.20.0
ffmpy==0.3.2
filelock==3.16.1
Flask==3.1.0
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.2.3
gensim==4.3.3
gradio==4.39.0
gradio_client==1.1.1
grpcio==1.65.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.2
humanfriendly==10.0
icecream==2.1.3
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
lightgbm==4.6.0
lightning==2.3.3
lightning-utilities==0.11.6
llvmlite==0.43.0
loguru==0.7.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mlxtend==0.23.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.17
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.1
nltk==3.9.1
node2vec==0.5.0
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open_clip_torch==2.30.0
openai==1.57.4
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optimum==1.23.1
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pathlib2==2.3.7.post1
pathos==0.3.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==4.3.6
pooch==1.8.2
pox==0.3.5
ppft==1.7.6.9
prometheus_client==0.20.0
prompt_toolkit==3.0.48
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==9.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.0.9
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytorch-crf==0.7.2
pytorch-lightning==2.3.3
pytorch-toolbelt==0.6.3
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.7.24
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rouge==1.0.1
rouge-chinese==1.0.3
rpds-py==0.19.1
ruff==0.5.5
s3transfer==0.10.4
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==3.3.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.5.0.post1
stack-data==0.6.3
stanza==1.10.1
starlette==0.37.2
supar==1.1.4
sympy==1.13.3
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
terminado==0.18.1
threadpoolctl==3.5.0
timm==1.0.11
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
torch==2.4.0
torchaudio==2.4.0
torchmetrics==1.4.0.post0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.44.2
triton==3.0.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.30.3
uvloop==0.21.0
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
Whoosh==2.7.4
widgetsnbextension==4.0.11
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.20.2
